+++ src/examples/dynamicHaDemo/src/app/dhawrapper/clCompAppMain.C
@@ -263,7 +263,7 @@ int main(int argc, char *argv[])
 
     clEoMyEoIocPortGet(&iocPort);
     
-    clprintf (CL_LOG_SEV_INFO, "Component [%s] : PID [%d]. Initializing\n", appName.value, mypid);
+    clprintf (CL_LOG_SEV_INFO, "Component [%s] : PID [%d]. Initializing\n", appName.value, (int)mypid);
     clprintf (CL_LOG_SEV_INFO, "   IOC Address             : 0x%x\n", clIocLocalAddressGet());
     clprintf (CL_LOG_SEV_INFO, "   IOC Port                : 0x%x\n", iocPort);
 
@@ -317,7 +317,7 @@ int main(int argc, char *argv[])
 errorexit:
 
     clprintf (CL_LOG_SEV_ERROR, "Component [%s] : PID [%d]. Initialization error [0x%x]\n",
-              appName.value, mypid, rc);
+              appName.value, (int)mypid, rc);
 
     return -1;
 }
@@ -334,7 +334,7 @@ void clCompAppTerminate(SaInvocationT invocation, const SaNameT *compName)
     SaAisErrorT rc = SA_AIS_OK;
 
     clprintf (CL_LOG_SEV_INFO, "Component [%s] : PID [%d]. Terminating\n",
-              compName->value, mypid);
+              compName->value, (int)mypid);
 
     /*
      * ---BEGIN_APPLICATION_CODE--- 
@@ -357,7 +357,7 @@ void clCompAppTerminate(SaInvocationT invocation, const SaNameT *compName)
     saAmfResponse(amfHandle, invocation, SA_AIS_OK);
 
     clprintf (CL_LOG_SEV_INFO, "Component [%s] : PID [%d]. Terminated\n",
-              compName->value, mypid);
+              compName->value, (int)mypid);
 
     unblockNow = CL_TRUE;
     
@@ -366,7 +366,7 @@ void clCompAppTerminate(SaInvocationT invocation, const SaNameT *compName)
 errorexit:
 
     clprintf (CL_LOG_SEV_ERROR, "Component [%s] : PID [%d]. Termination error [0x%x]\n",
-              compName->value, mypid, rc);
+              compName->value, (int)mypid, rc);
 
     return;
 }
@@ -484,7 +484,7 @@ void clCompAppAMFCSISet(SaInvocationT       invocation,
      */
 
     clprintf (CL_LOG_SEV_INFO, "Component [%s] : PID [%d]. CSI Set Received\n", 
-              compName->value, mypid);
+              compName->value, (int)mypid);
 
     clCompAppAMFPrintCSI(csiDescriptor, haState);
 
@@ -612,7 +612,7 @@ void clCompAppAMFCSIRemove(SaInvocationT  invocation,
                            SaAmfCSIFlagsT csiFlags)
 {
     clprintf (CL_LOG_SEV_INFO, "Component [%s] : PID [%d]. CSI Remove Received\n", 
-              compName->value, mypid);
+              compName->value, (int)mypid);
 
     clprintf (CL_LOG_SEV_INFO, "   CSI                     : %s\n", csiName->value);
     clprintf (CL_LOG_SEV_INFO, "   CSI Flags               : 0x%d\n", csiFlags);
